Refactor to move LogicalArchitectureModels to own module and creating an AbstractLogicalResourceInfo class
