+++ src/benchq/__init__.py
@@ -2,7 +2,3 @@
 # © Copyright 2022 Zapata Computing Inc.
 ################################################################################
 
-from .quantum_hardware_modeling.hardware_architecture_models import (
-    BASIC_ION_TRAP_ARCHITECTURE_MODEL,
-    BASIC_SC_ARCHITECTURE_MODEL,
-)
