+++ src/benchq/logical_architecture_modeling/basic_logical_architectures.py
@@ -1,0 +1,8 @@
+
+class LogicalArchitectureModel:
+    def __init__(self):
+        self._name = None
+
+    @property
+    def name(self):
+        return self._name
